pyproject.toml
[project]
name = "jax_tpu_embedding"
description = "JAX SparseCore API"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [{name = "JAX SC Authors", email="jax-sc-dev@google.com"}]
classifiers = [  # List of https://pypi.org/classifiers/
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
]
dependencies = [
    'jax',
    'absl-py',
    'flax',
    'numpy',
]

# `version` is set by setuptools to use `jax_tpu_embedding.sparsecore.__version__`.
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/jax-ml/jax_tpu_embedding"
repository = "https://github.com/jax-ml/jax_tpu_embedding"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Pseudo extension module to indicate to setuptools that the package
# does contain prebuilt extensions.
ext-modules = [{name = "jax_tpu_embedding.sparsecore.prebuilt", sources = []}]

[tool.setuptools.dynamic]
version = {attr = "jax_tpu_embedding.sparsecore.__version__"}

[tool.setuptools.package-data]
"*" = ["*.dylib", "*.so", "*.pyd"]
